Переделал роутинг через HOC - requireAuth
