+++ packages/client/src/app/router/router.tsx
@@ -14,6 +14,7 @@ import { ForumPage } from '../../pages/ForumPage'
 import { ForumThreadPage } from '../../pages/ForumThreadPage'
 import { ForumCreateThreadPage } from '../../pages/ForumCreateThreadPage'
 import { actionPaths } from '../../features/forums/types'
+import { RequireAuth } from '../../components/requireAuth/RequireAuth'
 
 export const router = createBrowserRouter([
   {
@@ -43,15 +44,27 @@ export const router = createBrowserRouter([
   },
   {
     path: RoutesEnum.Profile,
-    element: <ProfilePage />,
+    element: (
+      <RequireAuth>
+        <ProfilePage />
+      </RequireAuth>
+    ),
   },
   {
     path: RoutesEnum.Password,
-    element: <PasswordPage />,
+    element: (
+      <RequireAuth>
+        <PasswordPage />
+      </RequireAuth>
+    ),
   },
   {
     path: RoutesEnum.Forums,
-    element: <ForumsPage />,
+    element: (
+      <RequireAuth>
+        <ForumsPage />
+      </RequireAuth>
+    ),
     children: [
       {
         path: ':forum',
@@ -71,10 +84,18 @@ export const router = createBrowserRouter([
   },
   {
     path: RoutesEnum.Leaderboard,
-    element: <LeaderboardPage />,
+    element: (
+      <RequireAuth>
+        <LeaderboardPage />
+      </RequireAuth>
+    ),
   },
   {
     path: RoutesEnum.Game,
-    element: <GamePage />,
+    element: (
+      <RequireAuth>
+        <GamePage />
+      </RequireAuth>
+    ),
   },
 ])
