+++ packages/client/src/components/requireAuth/types.ts
@@ -1,0 +1,5 @@
+import { ReactElement } from 'react'
+
+export interface Children {
+  children: ReactElement
+}
